preserve thrown error on peer
This commit saves the error that causes a peer connection or initilaization to fail. This error can then be retrieved even after it has been thrown and caught/handled. It creates a global types class for types that are not specific to any particular module, which contains a type definition for internal xud errors.