enable add/remove pairs & currencies from grpc
This commit adds grpc calls for `addCurrency`, `addPair`, `removeCurrency`, and `removePair`. This allows the user to modify supported currencies and trading pairs without needing to edit the database manually.

It removes the `SwapProtocol` enum and instead tracks the payment channel network client used for swaps on a per-currency basis using a new `NetworkClient` enum. This is a breaking database change that will require the `currencies` and `pairs` tables to be recreated.

New service tests are added to test the various use cases for the newly added calls, along with corresponding cli commands.